Move override method into its own directory
